Add base query with pagination and limit

+++ src/app/dto/metaDTO.go
@@ -4,6 +4,6 @@ type MetaDTO struct {
 	Count      int `json:"count"`
 	Page       int `json:"page"`
 	TotalPages int `json:"total_pages"`
-	TotalCount int `json:"total_count"`
+	TotalCount int64 `json:"total_count"`
 	Limit      int `json:"limit"`
 }
